+++ src/test/java/com/pax/market/api/sdk/java/api/test/TerminalApiTest.java
@@ -576,8 +576,8 @@ public class TerminalApiTest {
     }
 
     @Test
-    public void testSearchTerminalTamperAlarm() {
-        Result<TerminalAlarmDTO> result = terminalApi.searchTerminalTamperAlarm(1, 10, null, null, null);
+    public void testSearchTamperedTerminal() {
+        Result<TerminalAlarmDTO> result = terminalApi.searchTamperedTerminal(1, 10, null, null, null);
         logger.debug("Result of search terminal tamper alarm: {}", result.toString());
         Assert.assertEquals(0, result.getBusinessCode());
     }